+++ src/Resources/contao/config/config.php
@@ -224,7 +224,7 @@ $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE']['contao']['core_form']['fil
  */
 $GLOBALS['NOTIFICATION_CENTER']['NOTIFICATION_TYPE']['smartgear']['ticket_creation'] = [
     'email_sender_address' => ['sg_owner_email'],
-    'recipients' => ['support_email'],
+    'recipients' => ['support_email', 'sg_owner_email'],
     'email_subject' => ['ticket_*'],
     'email_text' => ['ticket_*', 'sg_owner_name'],
     'email_html' => ['ticket_*', 'sg_owner_name'],